+++ packages/abi-gen/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.1.0 - _January 11, 2018_
 
+* Added CLI options for explicit specifying location of partials and main template
 * Fixed array typings with union types (#295)
 * Add event ABIs to context data passed to templates (#302)
 * Add constructor ABIs to context data passed to templates (#304)